refactor(additional resrouces): move to seperate directory
