+++ src/components/AdditionalMaterials/index.tsx
@@ -1,42 +1,64 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import { HelpCircle } from 'react-feather';
-import { Styles } from '../../constants/style';
-import { SbNavLink, SbNavlinkProps } from '../SbNavLink';
+import css from 'styled-jsx/css';
+import { Colors } from '../../constants/style';
+import { digitalLibraryLinks } from '../../constants/links';
+
+const linkStyle = css`
+  a {
+    color: ${Colors.sbBlueLighter};
+  }
+`;
 
 export interface AdditionalMaterialsProps {
-  links: SbNavlinkProps[];
+  target: string;
 }
 
-enum Content {
-  sampleItemsContent = 'Sample Items, Scoring Guides and Rubrics',
-  annotatedContent = 'Annotated Student Work',
-  digitalLibraryContent = 'Digital Library Resources'
-}
+export const AdditionalMaterials: React.SFC<AdditionalMaterialsProps> = ({ target }) => {
+  const grade = target.split('.')[1].split('G')[1];
+  const subject = target.split('.')[0] === 'E' ? 'ELA' : 'MATH';
+  const claim = target.split('.')[2].split('')[1];
+  const targ = target.split('.')[3].split('T')[1];
+  const dlLink = digitalLibraryLinks.get(target);
+  const materials = () => {
+    const includedMaterials = [
+      <Fragment>
+        <a
+          href={`https://sampleitems.smarterbalanced.org/BrowseItems?Claim=${subject}${claim}&Subject=${subject}&Grade=${grade}&Target=${targ}`}
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Sample Items, Scoring Guides and Rubrics
+        </a>
+        <style jsx>{linkStyle}</style>
+      </Fragment>
+    ];
+    const dlLink = digitalLibraryLinks.get(target);
+    if (dlLink) {
+      includedMaterials.push(
+        <Fragment>
+          <a href={dlLink} target="_blank" rel="noopener noreferrer">
+            Digital Library Resources
+          </a>
+          <style jsx>{linkStyle}</style>
+        </Fragment>
+      );
+    }
 
-export const AdditionalMaterials: React.SFC = (): JSX.Element => {
-  const buildContent = (content: string): JSX.Element => (
-    <div>
-      <HelpCircle />
-      <p>{content}</p>
+    return includedMaterials;
+  };
+
+  const materialsJsx = materials().map((m, i) => (
+    <div key={i}>
+      {m}
       <style jsx>{`
         div {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-        }
-        p {
-          font-family: ${Styles.sbSans};
-          padding: 0 10px;
+          padding: 5px;
+          margin-left: 10px;
         }
       `}</style>
     </div>
-  );
+  ));
 
-  return (
-    <div>
-      <SbNavLink url={''} content={buildContent(Content.sampleItemsContent)} />
-      <SbNavLink url={''} content={buildContent(Content.annotatedContent)} />
-      <SbNavLink url={''} content={buildContent(Content.digitalLibraryContent)} />
-    </div>
-  );
+  return <React.Fragment>{materialsJsx}</React.Fragment>;
 };
